feat: modified requested changes #112
